refactor: Restructure transactions
