+++ packages/cryptoplease/lib/bl/transactions/details/transaction_details_bloc.dart
