+++ packages/cryptoplease/lib/bl/transactions/details/transaction_details_event.dart
