+++ packages/cryptoplease/lib/bl/transactions/details/transaction_details_state.dart
