+++ packages/cryptoplease/lib/bl/transactions/item/transaction_item_state.dart
