+++ packages/cryptoplease/lib/presentation/screens/authenticated/transaction_details_screen/components/basic_transaction_information.dart
@@ -1,4 +1,4 @@
-import 'package:cryptoplease/bl/transaction_details/transaction_details_bloc.dart';
+import 'package:cryptoplease/bl/transactions/details/transaction_details_bloc.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/details_view.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/value_views.dart';
