+++ packages/cryptoplease/lib/presentation/screens/authenticated/transaction_details_screen/components/instruction_details_item.dart
@@ -1,5 +1,5 @@
 import 'package:cryptoplease/bl/tokens/token.dart';
-import 'package:cryptoplease/bl/transaction_details/transaction_details_bloc.dart';
+import 'package:cryptoplease/bl/transactions/details/transaction_details_bloc.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/memo_instruction_details.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/spl_token_instruction_details.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/system_instruction_details.dart';
