+++ packages/cryptoplease/lib/presentation/screens/authenticated/transaction_details_screen/components/transaction_details_view.dart
@@ -1,5 +1,5 @@
 import 'package:cryptoplease/bl/tokens/token.dart';
-import 'package:cryptoplease/bl/transaction_details/transaction_details_bloc.dart';
+import 'package:cryptoplease/bl/transactions/details/transaction_details_bloc.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/transaction_details.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/transaction_details_screen/components/transaction_details_loading_view.dart';
 import 'package:flutter/material.dart';
