+++ packages/cryptoplease/lib/presentation/screens/authenticated/transactions_screen/components/transaction_item.dart
@@ -1,8 +1,8 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:cryptoplease/bl/amount.dart';
 import 'package:cryptoplease/bl/tokens/token.dart';
-import 'package:cryptoplease/bl/transaction_item/transaction_item_bloc.dart';
-import 'package:cryptoplease/bl/transactions/transactions_bloc.dart';
+import 'package:cryptoplease/bl/transactions/item/transaction_item_bloc.dart';
+import 'package:cryptoplease/bl/transactions/list/transactions_bloc.dart';
 import 'package:cryptoplease/gen/assets.gen.dart';
 import 'package:cryptoplease/l10n/device_locale.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
