+++ packages/cryptoplease/lib/presentation/screens/authenticated/transactions_screen/transactions_screen.dart
@@ -1,7 +1,7 @@
 import 'package:cryptoplease/bl/accounts/account.dart';
 import 'package:cryptoplease/bl/tokens/token.dart';
-import 'package:cryptoplease/bl/transaction_item/transaction_item_bloc.dart';
-import 'package:cryptoplease/bl/transactions/transactions_bloc.dart';
+import 'package:cryptoplease/bl/transactions/item/transaction_item_bloc.dart';
+import 'package:cryptoplease/bl/transactions/list/transactions_bloc.dart';
 import 'package:cryptoplease/l10n/l10n.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/components/app_bar.dart';
 import 'package:cryptoplease/presentation/screens/authenticated/components/header_buttons.dart';
